Latest update
I think something strange has happened, because I tried to edit the class whilst another notebook was running.

I have majorly restructured cos it was a mess. All the code is in NetworksTerm Paper --> Networks_Term_Paper --> Code --> Python --> working_multiple_eq_flow. Data is in the same place.

In summary, running it for SiouxFalls overnight, it should give the results for shutting each and every link. Will need to think of a way to get all the things we need into the results file.

I tried running for the Berlin network, but it's throwing errors because of the file formats. If you have some time to investigate that would be amazing, call if any issues.

Ben