Pushing results file for SiouxFalls
File is in working_multiple_eq_flow/results.

Can't store large files on GitHub, so the pickles will be in the drive. The current pickle is wrong - I realised something late. Everything was fine other than: a) I hadn't appended the equilibrium flow when no link had been dropped and b) for link 0 it wasn't dropping. Note that 0 refers to the first line in the file because of python's indexing.

It doesn't matter though, you can still set up the functions to compute the results for each network.

Going to run the Berlin ones overnight but I'm pretty sure they won't finish.

Ben